Refacto UI : move services to be compliant with new UI Software architecture
