+++ ui/main/src/app/business/services/global-style.service.ts
