+++ ui/main/src/app/business/services/perimeters.service.ts
@@ -11,7 +11,7 @@ import {Observable, Subject} from 'rxjs';
 import {environment} from '@env/environment';
 import {map, takeUntil, tap} from 'rxjs/operators';
 import {Injectable} from '@angular/core';
-import {CachedCrudService} from '@ofServices/cached-crud-service';
+import {CachedCrudService} from 'app/business/services/cached-crud-service';
 import {Perimeter} from '@ofModel/perimeter.model';;
 import {OpfabLoggerService} from './logs/opfab-logger.service';
 import {AlertMessageService} from './alert-message.service';
