+++ ui/main/src/app/business/services/rrule-reminder/rrule-reminder.service.ts
@@ -11,7 +11,7 @@ import {Injectable} from '@angular/core';
 import {RRuleReminderList} from './rrule-reminderList';
 import {AcknowledgeService} from 'app/business/services/acknowledge.service';
 import {LightCardsStoreService} from 'app/business/services/lightcards/lightcards-store.service';
-import {SoundNotificationService} from '@ofServices/sound-notification.service';
+import {SoundNotificationService} from 'app/business/services/sound-notification.service';
 import {LogOption, OpfabLoggerService} from 'app/business/services/logs/opfab-logger.service';
 import {ServerResponseStatus} from 'app/business/server/serverResponse';
 import {CardService} from '../card.service';
