+++ ui/main/src/app/modules/admin/services/sharing.service.ts
@@ -11,8 +11,8 @@ import {Injectable, OnDestroy} from '@angular/core';
 import {EntitiesService} from 'app/business/services/entities.service';
 import {GroupsService} from 'app/business/services/groups.service';
 import {UserService} from 'app/business/services/user.service';
-import {CrudService} from '@ofServices/crud-service';
-import {CachedCrudService} from '@ofServices/cached-crud-service';
+import {CrudService} from 'app/business/services/crud-service';
+import {CachedCrudService} from 'app/business/services/cached-crud-service';
 import {Observable, ReplaySubject, Subject} from 'rxjs';
 import {takeUntil} from 'rxjs/operators';
 import {PerimetersService} from 'app/business/services/perimeters.service';
