+++ ui/main/src/app/modules/core/application-loading/app-loaded-in-another-tab/app-loaded-in-another-tab.component.ts
@@ -13,7 +13,7 @@ import {LogOption, OpfabLoggerService} from 'app/business/services/logs/opfab-lo
 import {UrlLockService} from './url-lock.service';
 import {UserService} from 'app/business/services/user.service';
 import {ApplicationLoadingStep} from '../application-loading-step';
-import {SoundNotificationService} from '@ofServices/sound-notification.service';
+import {SoundNotificationService} from 'app/business/services/sound-notification.service';
 import {OpfabEventStreamService} from 'app/business/services/opfabEventStream.service';
 
 /** This component checks if the url of opfab is already in use
