+++ ui/main/src/app/modules/feed/components/card-list/filters/feed-sort/feed-sort.component.ts
@@ -11,7 +11,7 @@ import {Component, Input, OnDestroy, OnInit} from '@angular/core';
 import {FormControl, FormGroup} from '@angular/forms';
 import {Subject} from 'rxjs';
 import {takeUntil} from 'rxjs/operators';
-import {UserPreferencesService} from '@ofServices/user-preference.service';
+import {UserPreferencesService} from 'app/business/services/user-preference.service';
 import {SortService} from 'app/business/services/lightcards/sort.service';
 
 @Component({
