+++ ui/main/src/app/modules/realtimeusers/realtimeusers.component.ts
@@ -11,7 +11,7 @@ import {Component, OnDestroy, OnInit} from '@angular/core';
 import {UserService} from 'app/business/services/user.service';
 import {RealTimeScreen} from '@ofModel/real-time-screens.model';
 import {FormControl, FormGroup} from '@angular/forms';
-import {UserPreferencesService} from '@ofServices/user-preference.service';
+import {UserPreferencesService} from 'app/business/services/user-preference.service';
 import {Utilities} from '../../business/common/utilities';
 import {MultiSelectConfig} from '@ofModel/multiselect.model';
 import {EntitiesService} from 'app/business/services/entities.service';
