+++ ui/main/src/app/modules/share/archives-logging-filters/archives-logging-filters.component.ts
@@ -22,7 +22,7 @@ import {MultiSelectOption} from '@ofModel/multiselect.model';
 import {MessageLevel} from '@ofModel/message.model';
 import moment from 'moment';
 import {Utilities} from 'app/business/common/utilities';
-import {UserPreferencesService} from '@ofServices/user-preference.service';
+import {UserPreferencesService} from 'app/business/services/user-preference.service';
 import {UserService} from 'app/business/services/user.service';
 import {PermissionEnum} from '@ofModel/permission.model';
 import {AlertMessageService} from 'app/business/services/alert-message.service';
