+++ ui/main/src/app/modules/share/light-card/light-card.component.ts
@@ -17,9 +17,9 @@ import {AppService, PageType} from '@ofServices/app.service';
 import {EntitiesService} from 'app/business/services/entities.service';
 import {ProcessesService} from 'app/business/services/processes.service';
 import {DisplayContext} from '@ofModel/templateGateway.model';
-import {GroupedCardsService} from '@ofServices/grouped-cards.service';
+import {GroupedCardsService} from 'app/business/services/grouped-cards.service';
 import {TypeOfStateEnum} from '@ofModel/processes.model';
-import {SoundNotificationService} from '@ofServices/sound-notification.service';
+import {SoundNotificationService} from 'app/business/services/sound-notification.service';
 import {DateTimeFormatterService} from 'app/business/services/date-time-formatter.service';
 import {MapService} from 'app/business/services/map.service';
 import {TranslateService} from '@ngx-translate/core';
