+++ ui/main/src/app/modules/share/template-rendering/template-rendering.component.ts
@@ -32,7 +32,7 @@ import {User} from '@ofModel/user.model';
 import {OpfabLoggerService} from 'app/business/services/logs/opfab-logger.service';
 import {DisplayContext} from '@ofModel/templateGateway.model';
 import {TemplateCssService} from 'app/business/services/template-css.service';
-import {GlobalStyleService} from '@ofServices/global-style.service';
+import {GlobalStyleService} from 'app/business/services/global-style.service';
 import {CurrentUserStore} from 'app/business/store/current-user.store';
 import {UserService} from 'app/business/services/user.service';
